Restructures Unit-Tests into tests-Folder
