+++ packages/place-mark-api/tests/unit-tests/controllers/api/user-api-controller.spec.ts
@@ -1,10 +1,10 @@
 import { Server, ServerInjectResponse } from "@hapi/hapi";
 import { assert } from "chai";
-import { ContainerMock, testConfig } from "../controller-test-setup.spec.js";
-import { createServer$ } from "../../server.js";
-import { IPaginatedListRequest, IPaginatedListResponse, IUserReadOnlyDto, IUserRepository } from "../../core/index.js";
-import { IValidationResult } from "./user-api-controller.js";
-import { toQueryString, QueryParams } from "../../../tests/utils.js";
+import { ContainerMock, testConfig } from "../test-setup.js";
+import { createServer$ } from "../../../../app/server.js";
+import { IPaginatedListRequest, IPaginatedListResponse, IUserReadOnlyDto, IUserRepository } from "../../../../app/core/index.js";
+import { IValidationResult } from "../../../../app/controllers/index.js";
+import { toQueryString, QueryParams } from "../../../utils.js";
 
 suite("UserApiController Unit-Tests", () => {
   let server: Server;
